Initial commit
At this point the mod can change the name of train stop using the red circuit wire.
All train stops built while this mod is not loaded, is not affected by this mod.
All signals used for other control/indication behavior of the train stop is exclued from the name.